Avoid ToolConfig TimeSpan Overflow
The PredictRuntime implementations may attempt to generate a TimeSpan
object that exceeds TimeSpan.MaxValue in total length.  I've added a
wrapper to ToolConfig and changed the interface implemented by
OnionToolConfig and GpgToolConfig so that they can not cause this
exception.

The compromise is that when the timespan is unrealistically large the
value will be incorrect, however since that requires a timespan of over
10,000,000 days it is highly unlikely the user will survive long enough
to notice.